+++ src/Facades/MediaManager.php
@@ -12,6 +12,9 @@ use Spatie\MediaLibrary\MediaCollections\Models\Media;
  * @method static Media add(string $path, ?string $folder = null, string $disk = 'public', ?Model $vault = null, ?string $fileName = null, bool $unique = false, bool $preservingOriginal = true)
  * @method static Media addFromFile(string $absolutePath, ?string $folder = null, ?Model $vault = null, ?string $fileName = null, bool $unique = false, bool $preservingOriginal = true)
  * @method static string url(string $path, ?string $folder = null, string $disk = 'public', ?Model $vault = null, ?string $fileName = null, bool $unique = false, bool $preservingOriginal = true)
+ * @method static Media|null find(string $path, ?string $folder = null, ?Model $vault = null)
+ * @method static Media findOrFail(string $path, ?string $folder = null, ?Model $vault = null)
+ * @method static string|null findUrl(string $path, ?string $folder = null, ?Model $vault = null, string $conversion = '')
  * @method static MediaFolder|null folder(string $folder, ?Model $vault = null)
  * @method static MediaFolder|null findFolder(string $folder, ?Model $vault = null)
  *